All test cases for Teacher Contact Management
