--- TimetableManagementSystem/src/main/java/admin_mgmt/tt_adminTest.java
@@ -1,14 +1,0 @@
-package admin_mgmt;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.bson.conversions.Bson;
-import org.junit.jupiter.api.Test;
-
-import com.mongodb.client.model.Filters;
-
-class tt_adminTest {
-
-
-
-}
