+++ TimetableManagementSystem/src/main/java/teacher/contactDatabase.java
@@ -1,4 +1,4 @@
-package teacher_personalmgnt;
+package teacher;
 
 public class contactDatabase {
